Backend/chatpybackend/requirements.txt
flask
pandas
requests
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
Flask

